Moved service resource type filter to base
